signTx: update for stax
