+++ src/signTx.c
@@ -7,7 +7,6 @@
 #include "addressUtilsShelley.h"
 #include "uiHelpers.h"
 #include "signTxUtils.h"
-#include "uiScreens.h"
 #include "txHashBuilder.h"
 #include "textUtils.h"
 #include "hexUtils.h"
@@ -15,6 +14,7 @@
 #include "messageSigning.h"
 #include "bufView.h"
 #include "securityPolicy.h"
+#include "signTx_ui.h"
 
 static ins_sign_tx_context_t* ctx = &(instructionState.signTxContext);
 
@@ -58,7 +58,7 @@ static inline void initTxWitnessCtx()
 }
 
 // advances the stage of the main state machine
-static inline void advanceStage()
+void tx_advanceStage()
 {
 	TRACE("Advancing sign tx stage from: %d", ctx->stage);
 
@@ -292,7 +292,7 @@ static inline void advanceStage()
 		break;
 
 	case SIGN_STAGE_NONE:
-		// advanceStage() not supposed to be called after tx processing is finished
+		// tx_advanceStage() not supposed to be called after tx processing is finished
 		ASSERT(false);
 
 	default:
@@ -318,7 +318,7 @@ static inline void advanceCertificatesStateIfAppropriate()
 		BODY_CTX->currentCertificate++;
 
 		if (BODY_CTX->currentCertificate == ctx->numCertificates) {
-			advanceStage();
+			tx_advanceStage();
 		}
 	}
 	break;
@@ -350,7 +350,7 @@ static inline void checkForFinishedSubmachines()
 
 			BODY_CTX->currentOutput++;
 			if (BODY_CTX->currentOutput == ctx->numOutputs) {
-				advanceStage();
+				tx_advanceStage();
 			}
 		}
 		break;
@@ -375,7 +375,7 @@ static inline void checkForFinishedSubmachines()
 			STATIC_ASSERT(SIZEOF(AUX_DATA_CTX->stageContext.cvote_registration_subctx.auxDataHash) == AUX_DATA_HASH_LENGTH, "Wrong auxiliary data hash length");
 			memmove(ctx->auxDataHash, AUX_DATA_CTX->stageContext.cvote_registration_subctx.auxDataHash, AUX_DATA_HASH_LENGTH);
 
-			advanceStage();
+			tx_advanceStage();
 		}
 		break;
 
@@ -384,7 +384,7 @@ static inline void checkForFinishedSubmachines()
 			TRACE();
 			ctx->stage = SIGN_STAGE_BODY_MINT;
 			BODY_CTX->mintReceived = true;
-			advanceStage();
+			tx_advanceStage();
 		}
 
 	case SIGN_STAGE_BODY_COLLATERAL_OUTPUT_SUBMACHINE:
@@ -392,7 +392,7 @@ static inline void checkForFinishedSubmachines()
 			TRACE();
 			ctx->stage = SIGN_STAGE_BODY_COLLATERAL_OUTPUT;
 			BODY_CTX->collateralOutputReceived = true;
-			advanceStage();
+			tx_advanceStage();
 		}
 		break;
 
@@ -409,119 +409,6 @@ static inline void CHECK_STAGE(sign_tx_stage_t expected)
 }
 
 // ============================== INIT ==============================
-
-enum {
-	HANDLE_INIT_STEP_PROMPT_SIGNINGMODE = 100,
-	HANDLE_INIT_STEP_DISPLAY_NETWORK_DETAILS,
-	HANDLE_INIT_STEP_SCRIPT_RUNNING_WARNING,
-	HANDLE_INIT_STEP_NO_COLLATERAL_WARNING,
-	HANDLE_INIT_STEP_UNKNOWN_COLLATERAL_WARNING,
-	HANDLE_INIT_STEP_NO_SCRIPT_DATA_HASH_WARNING,
-	HANDLE_INIT_STEP_RESPOND,
-	HANDLE_INIT_STEP_INVALID,
-} ;
-
-static const char* _newTxLine1(sign_tx_signingmode_t txSigningMode)
-{
-	switch (txSigningMode) {
-
-	case SIGN_TX_SIGNINGMODE_ORDINARY_TX:
-		return "New ordinary";
-
-	case SIGN_TX_SIGNINGMODE_POOL_REGISTRATION_OWNER:
-		return "New pool owner";
-
-	case SIGN_TX_SIGNINGMODE_POOL_REGISTRATION_OPERATOR:
-		return "New pool operator";
-
-	case SIGN_TX_SIGNINGMODE_MULTISIG_TX:
-		return "New multisig";
-
-	case SIGN_TX_SIGNINGMODE_PLUTUS_TX:
-		return "New Plutus";
-
-	default:
-		ASSERT(false);
-	}
-}
-
-static void signTx_handleInit_ui_runStep()
-{
-	TRACE("UI step %d", ctx->ui_step);
-	TRACE_STACK_USAGE();
-	ui_callback_fn_t* this_fn = signTx_handleInit_ui_runStep;
-
-	UI_STEP_BEGIN(ctx->ui_step, this_fn);
-
-	UI_STEP(HANDLE_INIT_STEP_PROMPT_SIGNINGMODE) {
-		ui_displayPrompt(
-		        _newTxLine1(ctx->commonTxData.txSigningMode),
-		        "transaction?",
-		        this_fn,
-		        respond_with_user_reject
-		);
-	}
-
-	UI_STEP(HANDLE_INIT_STEP_DISPLAY_NETWORK_DETAILS) {
-		const bool isNetworkIdVerifiable = isTxNetworkIdVerifiable(
-		        ctx->includeNetworkId,
-		        ctx->numOutputs, ctx->numWithdrawals,
-		        ctx->commonTxData.txSigningMode
-		                                   );
-		if (isNetworkIdVerifiable) {
-			if (isNetworkUsual(ctx->commonTxData.networkId, ctx->commonTxData.protocolMagic)) {
-				// no need to display the network details
-				UI_STEP_JUMP(HANDLE_INIT_STEP_SCRIPT_RUNNING_WARNING);
-			}
-			ui_displayNetworkParamsScreen(
-			        "Network details",
-			        ctx->commonTxData.networkId, ctx->commonTxData.protocolMagic,
-			        this_fn
-			);
-		} else {
-			// technically, no pool reg. certificate as well, but the UI message would be too long
-			ui_displayPaginatedText(
-			        "Warning:",
-			        "cannot verify network id: no outputs or withdrawals",
-			        this_fn
-			);
-		}
-	}
-
-	UI_STEP(HANDLE_INIT_STEP_SCRIPT_RUNNING_WARNING) {
-		if (!needsRunningScriptWarning(ctx->numCollateralInputs)) {
-			UI_STEP_JUMP(HANDLE_INIT_STEP_NO_COLLATERAL_WARNING);
-		}
-		ui_displayPaginatedText("WARNING:", "Plutus script will be evaluated", this_fn);
-	}
-
-	UI_STEP(HANDLE_INIT_STEP_NO_COLLATERAL_WARNING) {
-		if (!needsMissingCollateralWarning(ctx->commonTxData.txSigningMode, ctx->numCollateralInputs)) {
-			UI_STEP_JUMP(HANDLE_INIT_STEP_NO_SCRIPT_DATA_HASH_WARNING);
-		}
-		ui_displayPaginatedText("WARNING:", "No collateral given for Plutus transaction", this_fn);
-	}
-
-	UI_STEP(HANDLE_INIT_STEP_UNKNOWN_COLLATERAL_WARNING) {
-		if (!needsUnknownCollateralWarning(ctx->commonTxData.txSigningMode, ctx->includeTotalCollateral)) {
-			UI_STEP_JUMP(HANDLE_INIT_STEP_NO_SCRIPT_DATA_HASH_WARNING);
-		}
-		ui_displayPaginatedText("WARNING:", "Unknown collateral amount", this_fn);
-	}
-
-	UI_STEP(HANDLE_INIT_STEP_NO_SCRIPT_DATA_HASH_WARNING) {
-		if (!needsMissingScriptDataHashWarning(ctx->commonTxData.txSigningMode, ctx->includeScriptDataHash)) {
-			UI_STEP_JUMP(HANDLE_INIT_STEP_RESPOND);
-		}
-		ui_displayPaginatedText("WARNING:", "No script data given for Plutus transaction", this_fn);
-	}
-
-	UI_STEP(HANDLE_INIT_STEP_RESPOND) {
-		respondSuccessEmptyMsg();
-		advanceStage();
-	}
-	UI_STEP_END(HANDLE_INIT_STEP_INVALID);
-}
 
 __noinline_due_to_stack__
 static void signTx_handleInitAPDU(uint8_t p2, const uint8_t* wireDataBuffer, size_t wireDataSize)
@@ -695,66 +582,6 @@ static void signTx_handleInitAPDU(uint8_t p2, const uint8_t* wireDataBuffer, siz
 
 // ============================== AUXILIARY DATA ==============================
 
-enum {
-	HANDLE_AUX_DATA_ARBITRARY_HASH_STEP_DISPLAY = 800,
-	HANDLE_AUX_DATA_ARBITRARY_HASH_STEP_RESPOND,
-	HANDLE_AUX_DATA_ARBITRARY_HASH_STEP_INVALID,
-};
-
-static void signTx_handleAuxDataArbitraryHash_ui_runStep()
-{
-	TRACE("UI step %d", ctx->ui_step);
-	TRACE_STACK_USAGE();
-	ui_callback_fn_t* this_fn = signTx_handleAuxDataArbitraryHash_ui_runStep;
-
-	UI_STEP_BEGIN(ctx->ui_step, this_fn);
-
-	UI_STEP(HANDLE_AUX_DATA_ARBITRARY_HASH_STEP_DISPLAY) {
-		ui_displayHexBufferScreen(
-		        "Auxiliary data hash",
-		        ctx->auxDataHash,
-		        SIZEOF(ctx->auxDataHash),
-		        this_fn
-		);
-	}
-	UI_STEP(HANDLE_AUX_DATA_ARBITRARY_HASH_STEP_RESPOND) {
-		respondSuccessEmptyMsg();
-		advanceStage();
-	}
-	UI_STEP_END(HANDLE_AUX_DATA_ARBITRARY_HASH_STEP_INVALID);
-}
-
-
-enum {
-	HANDLE_AUX_DATA_CVOTE_REGISTRATION_STEP_DISPLAY = 850,
-	HANDLE_AUX_DATA_CVOTE_REGISTRATION_STEP_RESPOND,
-	HANDLE_AUX_DATA_CVOTE_REGISTRATION_STEP_INVALID,
-};
-
-static void signTx_handleAuxDataCVoteRegistration_ui_runStep()
-{
-	TRACE("UI step %d", ctx->ui_step);
-	TRACE_STACK_USAGE();
-	ui_callback_fn_t* this_fn = signTx_handleAuxDataCVoteRegistration_ui_runStep;
-
-	UI_STEP_BEGIN(ctx->ui_step, this_fn);
-
-	UI_STEP(HANDLE_AUX_DATA_CVOTE_REGISTRATION_STEP_DISPLAY) {
-		ui_displayPrompt(
-		        "Register vote",
-		        "key (CIP-36)?",
-		        this_fn,
-		        respond_with_user_reject
-		);
-	}
-	UI_STEP(HANDLE_AUX_DATA_CVOTE_REGISTRATION_STEP_RESPOND) {
-		respondSuccessEmptyMsg();
-		signTxCVoteRegistration_init();
-		ctx->stage = SIGN_STAGE_AUX_DATA_CVOTE_REGISTRATION_SUBMACHINE;
-	}
-	UI_STEP_END(HANDLE_AUX_DATA_CVOTE_REGISTRATION_STEP_INVALID);
-}
-
 __noinline_due_to_stack__
 static void signTx_handleAuxDataAPDU(uint8_t p2, const uint8_t* wireDataBuffer, size_t wireDataSize)
 {
@@ -843,32 +670,6 @@ static void signTx_handleAuxDataAPDU(uint8_t p2, const uint8_t* wireDataBuffer, 
 
 // ============================== INPUTS ==============================
 
-enum {
-	HANDLE_INPUT_STEP_DISPLAY = 200,
-	HANDLE_INPUT_STEP_RESPOND,
-	HANDLE_INPUT_STEP_INVALID,
-};
-
-static void signTx_handleInput_ui_runStep()
-{
-	TRACE("UI step %d", ctx->ui_step);
-	ui_callback_fn_t* this_fn = signTx_handleInput_ui_runStep;
-
-	UI_STEP_BEGIN(ctx->ui_step, this_fn);
-
-	UI_STEP(HANDLE_INPUT_STEP_DISPLAY) {
-		ui_displayInputScreen(&BODY_CTX->stageData.input, this_fn);
-	}
-
-	UI_STEP(HANDLE_INPUT_STEP_RESPOND) {
-		respondSuccessEmptyMsg();
-
-		ASSERT(ctx->ui_advanceState != NULL);
-		ctx->ui_advanceState();
-	}
-	UI_STEP_END(HANDLE_INPUT_STEP_INVALID);
-}
-
 // Advance stage to the next input
 static void ui_advanceState_input()
 {
@@ -876,7 +677,7 @@ static void ui_advanceState_input()
 	BODY_CTX->currentInput++;
 
 	if (BODY_CTX->currentInput == ctx->numInputs) {
-		advanceStage();
+		tx_advanceStage();
 	}
 }
 
@@ -985,31 +786,6 @@ static void signTx_handleOutputAPDU(uint8_t p2, const uint8_t* wireDataBuffer, s
 
 // ============================== FEE ==============================
 
-enum {
-	HANDLE_FEE_STEP_DISPLAY = 400,
-	HANDLE_FEE_STEP_RESPOND,
-	HANDLE_FEE_STEP_INVALID,
-};
-
-static void signTx_handleFee_ui_runStep()
-{
-	TRACE("UI step %d", ctx->ui_step);
-	ui_callback_fn_t* this_fn = signTx_handleFee_ui_runStep;
-
-	TRACE_ADA_AMOUNT("fee ", BODY_CTX->stageData.fee);
-
-	UI_STEP_BEGIN(ctx->ui_step, this_fn);
-
-	UI_STEP(HANDLE_FEE_STEP_DISPLAY) {
-		ui_displayAdaAmountScreen("Transaction fee", BODY_CTX->stageData.fee, this_fn);
-	}
-	UI_STEP(HANDLE_FEE_STEP_RESPOND) {
-		respondSuccessEmptyMsg();
-		advanceStage();
-	}
-	UI_STEP_END(HANDLE_FEE_STEP_INVALID);
-}
-
 __noinline_due_to_stack__
 static void signTx_handleFeeAPDU(uint8_t p2, const uint8_t* wireDataBuffer, size_t wireDataSize)
 {
@@ -1056,34 +832,6 @@ static void signTx_handleFeeAPDU(uint8_t p2, const uint8_t* wireDataBuffer, size
 
 
 // ============================== TTL ==============================
-
-enum {
-	HANDLE_TTL_STEP_DISPLAY = 500,
-	HANDLE_TTL_STEP_RESPOND,
-	HANDLE_TTL_STEP_INVALID,
-};
-
-static void signTx_handleTtl_ui_runStep()
-{
-	TRACE("UI step %d", ctx->ui_step);
-	ui_callback_fn_t* this_fn = signTx_handleTtl_ui_runStep;
-
-	UI_STEP_BEGIN(ctx->ui_step, this_fn);
-
-	UI_STEP(HANDLE_TTL_STEP_DISPLAY) {
-		ui_displayValidityBoundaryScreen(
-		        "Transaction TTL",
-		        BODY_CTX->stageData.ttl,
-		        ctx->commonTxData.networkId, ctx->commonTxData.protocolMagic,
-		        this_fn
-		);
-	}
-	UI_STEP(HANDLE_TTL_STEP_RESPOND) {
-		respondSuccessEmptyMsg();
-		advanceStage();
-	}
-	UI_STEP_END(HANDLE_TTL_STEP_INVALID);
-}
 
 __noinline_due_to_stack__
 static void signTx_handleTtlAPDU(uint8_t p2, const uint8_t* wireDataBuffer, size_t wireDataSize)
@@ -1132,173 +880,6 @@ static void signTx_handleTtlAPDU(uint8_t p2, const uint8_t* wireDataBuffer, size
 
 
 // ============================== CERTIFICATES ==============================
-
-enum {
-	HANDLE_CERTIFICATE_STEP_DISPLAY_OPERATION = 600,
-	HANDLE_CERTIFICATE_STEP_DISPLAY_STAKING_KEY,
-	HANDLE_CERTIFICATE_STEP_CONFIRM,
-	HANDLE_CERTIFICATE_STEP_RESPOND,
-	HANDLE_CERTIFICATE_STEP_INVALID,
-};
-
-static void signTx_handleCertificate_ui_runStep()
-{
-	TRACE("UI step %d", ctx->ui_step);
-	ui_callback_fn_t* this_fn = signTx_handleCertificate_ui_runStep;
-
-	UI_STEP_BEGIN(ctx->ui_step, this_fn);
-
-	UI_STEP(HANDLE_CERTIFICATE_STEP_DISPLAY_OPERATION) {
-		switch (BODY_CTX->stageData.certificate.type) {
-		case CERTIFICATE_TYPE_STAKE_REGISTRATION:
-			ui_displayPaginatedText(
-			        "Register",
-			        "staking key",
-			        this_fn
-			);
-			break;
-
-		case CERTIFICATE_TYPE_STAKE_DEREGISTRATION:
-			ui_displayPaginatedText(
-			        "Deregister",
-			        "staking key",
-			        this_fn
-			);
-			break;
-
-		case CERTIFICATE_TYPE_STAKE_DELEGATION:
-			ui_displayBech32Screen(
-			        "Delegate stake to",
-			        "pool",
-			        BODY_CTX->stageData.certificate.poolKeyHash, SIZEOF(BODY_CTX->stageData.certificate.poolKeyHash),
-			        this_fn
-			);
-			break;
-
-		default:
-			// includes CERTIFICATE_TYPE_STAKE_POOL_REGISTRATION
-			// and CERTIFICATE_TYPE_STAKE_POOL_RETIREMENT
-			// which have separate UI; this handler must not be used
-			ASSERT(false);
-		}
-	}
-	UI_STEP(HANDLE_CERTIFICATE_STEP_DISPLAY_STAKING_KEY) {
-		switch (BODY_CTX->stageData.certificate.stakeCredential.type) {
-		case STAKE_CREDENTIAL_KEY_PATH:
-			ui_displayPathScreen(
-			        "Staking key",
-			        &BODY_CTX->stageData.certificate.stakeCredential.keyPath,
-			        this_fn
-			);
-			break;
-		case STAKE_CREDENTIAL_KEY_HASH:
-			ui_displayBech32Screen(
-			        "Staking key hash",
-			        "stake_vkh",
-			        BODY_CTX->stageData.certificate.stakeCredential.keyHash,
-			        SIZEOF(BODY_CTX->stageData.certificate.stakeCredential.keyHash),
-			        this_fn
-			);
-			break;
-		case STAKE_CREDENTIAL_SCRIPT_HASH:
-			ui_displayBech32Screen(
-			        "Staking script hash",
-			        "script",
-			        BODY_CTX->stageData.certificate.stakeCredential.scriptHash,
-			        SIZEOF(BODY_CTX->stageData.certificate.stakeCredential.scriptHash),
-			        this_fn
-			);
-			break;
-		default:
-			ASSERT(false);
-			break;
-		}
-	}
-	UI_STEP(HANDLE_CERTIFICATE_STEP_CONFIRM) {
-		char description[50] = {0};
-		explicit_bzero(description, SIZEOF(description));
-
-		switch (BODY_CTX->stageData.certificate.type) {
-		case CERTIFICATE_TYPE_STAKE_REGISTRATION:
-			snprintf(description, SIZEOF(description), "registration?");
-			break;
-
-		case CERTIFICATE_TYPE_STAKE_DEREGISTRATION:
-			snprintf(description, SIZEOF(description), "deregistration?");
-			break;
-
-		case CERTIFICATE_TYPE_STAKE_DELEGATION:
-			snprintf(description, SIZEOF(description), "delegation?");
-			break;
-
-		default:
-			ASSERT(false);
-		}
-		// make sure all the information is displayed to the user
-		ASSERT(strlen(description) + 1 < SIZEOF(description));
-
-		ui_displayPrompt(
-		        "Confirm",
-		        description,
-		        this_fn,
-		        respond_with_user_reject
-		);
-	}
-	UI_STEP(HANDLE_CERTIFICATE_STEP_RESPOND) {
-		respondSuccessEmptyMsg();
-
-		advanceCertificatesStateIfAppropriate();
-	}
-	UI_STEP_END(HANDLE_CERTIFICATE_STEP_INVALID);
-}
-
-enum {
-	HANDLE_CERTIFICATE_POOL_RETIREMENT_STEP_DISPLAY_OPERATION = 650,
-	HANDLE_CERTIFICATE_POOL_RETIREMENT_STEP_DISPLAY_EPOCH,
-	HANDLE_CERTIFICATE_POOL_RETIREMENT_STEP_CONFIRM,
-	HANDLE_CERTIFICATE_POOL_RETIREMENT_STEP_RESPOND,
-	HANDLE_CERTIFICATE_POOL_RETIREMENT_STEP_INVALID,
-};
-
-static void signTx_handleCertificatePoolRetirement_ui_runStep()
-{
-	TRACE("UI step %d", ctx->ui_step);
-	ASSERT(BODY_CTX->stageData.certificate.type == CERTIFICATE_TYPE_STAKE_POOL_RETIREMENT);
-
-	ui_callback_fn_t* this_fn = signTx_handleCertificatePoolRetirement_ui_runStep;
-
-	UI_STEP_BEGIN(ctx->ui_step, this_fn);
-
-	UI_STEP(HANDLE_CERTIFICATE_POOL_RETIREMENT_STEP_DISPLAY_OPERATION) {
-		ui_displayBech32Screen(
-		        "Retire stake pool",
-		        "pool",
-		        BODY_CTX->stageData.certificate.poolKeyHash, SIZEOF(BODY_CTX->stageData.certificate.poolKeyHash),
-		        this_fn
-		);
-	}
-	UI_STEP(HANDLE_CERTIFICATE_POOL_RETIREMENT_STEP_DISPLAY_EPOCH) {
-		ui_displayUint64Screen(
-		        "at the start of epoch",
-		        BODY_CTX->stageData.certificate.epoch,
-		        this_fn
-		);
-	}
-	UI_STEP(HANDLE_CERTIFICATE_POOL_RETIREMENT_STEP_CONFIRM) {
-		ui_displayPrompt(
-		        "Confirm",
-		        "pool retirement",
-		        this_fn,
-		        respond_with_user_reject
-		);
-	}
-	UI_STEP(HANDLE_CERTIFICATE_POOL_RETIREMENT_STEP_RESPOND) {
-		respondSuccessEmptyMsg();
-
-		advanceCertificatesStateIfAppropriate();
-	}
-	UI_STEP_END(HANDLE_CERTIFICATE_POOL_RETIREMENT_STEP_INVALID);
-}
 
 static void _parsePathSpec(read_view_t* view, bip44_path_t* pathSpec)
 {
@@ -1567,76 +1148,6 @@ static void signTx_handleCertificateAPDU(uint8_t p2, const uint8_t* wireDataBuff
 
 // ============================== WITHDRAWALS ==============================
 
-enum {
-	HANDLE_WITHDRAWAL_STEP_DISPLAY_AMOUNT = 700,
-	HANDLE_WITHDRAWAL_STEP_DISPLAY_PATH,
-	HANDLE_WITHDRAWAL_STEP_RESPOND,
-	HANDLE_WITHDRAWAL_STEP_INVALID,
-};
-
-static void signTx_handleWithdrawal_ui_runStep()
-{
-	TRACE("UI step %d", ctx->ui_step);
-	TRACE_STACK_USAGE();
-	ui_callback_fn_t* this_fn = signTx_handleWithdrawal_ui_runStep;
-
-	UI_STEP_BEGIN(ctx->ui_step, this_fn);
-
-	UI_STEP(HANDLE_WITHDRAWAL_STEP_DISPLAY_AMOUNT) {
-		ui_displayAdaAmountScreen("Withdrawing rewards", BODY_CTX->stageData.withdrawal.amount, this_fn);
-	}
-	UI_STEP(HANDLE_WITHDRAWAL_STEP_DISPLAY_PATH) {
-		reward_account_t rewardAccount;
-		switch (BODY_CTX->stageData.withdrawal.stakeCredential.type) {
-		case STAKE_CREDENTIAL_KEY_PATH: {
-			rewardAccount.keyReferenceType = KEY_REFERENCE_PATH;
-			rewardAccount.path = BODY_CTX->stageData.withdrawal.stakeCredential.keyPath;
-			break;
-		}
-		case STAKE_CREDENTIAL_KEY_HASH: {
-			rewardAccount.keyReferenceType = KEY_REFERENCE_HASH;
-			constructRewardAddressFromHash(
-			        ctx->commonTxData.networkId,
-			        REWARD_HASH_SOURCE_KEY,
-			        BODY_CTX->stageData.withdrawal.stakeCredential.keyHash,
-			        SIZEOF(BODY_CTX->stageData.withdrawal.stakeCredential.keyHash),
-			        rewardAccount.hashBuffer,
-			        SIZEOF(rewardAccount.hashBuffer)
-			);
-			break;
-		}
-		case STAKE_CREDENTIAL_SCRIPT_HASH: {
-			rewardAccount.keyReferenceType = KEY_REFERENCE_HASH;
-			constructRewardAddressFromHash(
-			        ctx->commonTxData.networkId,
-			        REWARD_HASH_SOURCE_SCRIPT,
-			        BODY_CTX->stageData.withdrawal.stakeCredential.scriptHash,
-			        SIZEOF(BODY_CTX->stageData.withdrawal.stakeCredential.scriptHash),
-			        rewardAccount.hashBuffer,
-			        SIZEOF(rewardAccount.hashBuffer)
-			);
-			break;
-		}
-		default:
-			ASSERT(false);
-			break;
-		}
-		ui_displayRewardAccountScreen(&rewardAccount, ctx->commonTxData.networkId, this_fn);
-	}
-	UI_STEP(HANDLE_WITHDRAWAL_STEP_RESPOND) {
-		respondSuccessEmptyMsg();
-
-		// Advance stage to the next withdrawal
-		ASSERT(BODY_CTX->currentWithdrawal < ctx->numWithdrawals);
-		BODY_CTX->currentWithdrawal++;
-
-		if (BODY_CTX->currentWithdrawal == ctx->numWithdrawals) {
-			advanceStage();
-		}
-	}
-	UI_STEP_END(HANDLE_WITHDRAWAL_STEP_INVALID);
-}
-
 __noinline_due_to_stack__
 static void _addWithdrawalToTxHash(bool validateCanonicalOrdering)
 {
@@ -1755,34 +1266,6 @@ static void signTx_handleWithdrawalAPDU(uint8_t p2, const uint8_t* wireDataBuffe
 
 // ============================== VALIDITY INTERVAL START ==============================
 
-enum {
-	HANDLE_VALIDITY_INTERVAL_START_STEP_DISPLAY = 900,
-	HANDLE_VALIDITY_INTERVAL_START_STEP_RESPOND,
-	HANDLE_VALIDITY_INTERVAL_START_STEP_INVALID,
-};
-
-static void signTx_handleValidityInterval_ui_runStep()
-{
-	TRACE("UI step %d", ctx->ui_step);
-	ui_callback_fn_t* this_fn = signTx_handleValidityInterval_ui_runStep;
-
-	UI_STEP_BEGIN(ctx->ui_step, this_fn);
-
-	UI_STEP(HANDLE_VALIDITY_INTERVAL_START_STEP_DISPLAY) {
-		ui_displayValidityBoundaryScreen(
-		        "Validity interval start",
-		        BODY_CTX->stageData.validityIntervalStart,
-		        ctx->commonTxData.networkId, ctx->commonTxData.protocolMagic,
-		        this_fn
-		);
-	}
-	UI_STEP(HANDLE_VALIDITY_INTERVAL_START_STEP_RESPOND) {
-		respondSuccessEmptyMsg();
-		advanceStage();
-	}
-	UI_STEP_END(HANDLE_VALIDITY_INTERVAL_START_STEP_INVALID);
-}
-
 static void signTx_handleValidityIntervalStartAPDU(uint8_t p2, const uint8_t* wireDataBuffer, size_t wireDataSize)
 {
 	{
@@ -1854,35 +1337,6 @@ static void signTx_handleMintAPDU(uint8_t p2, const uint8_t* wireDataBuffer, siz
 
 // ========================= SCRIPT DATA HASH ==========================
 
-
-enum {
-	HANDLE_SCRIPT_DATA_HASH_STEP_DISPLAY = 1200,
-	HANDLE_SCRIPT_DATA_HASH_STEP_RESPOND,
-	HANDLE_SCRIPT_DATA_HASH_STEP_INVALID,
-};
-
-static void signTx_handleScriptDataHash_ui_runStep()
-{
-	TRACE("UI step %d", ctx->ui_step);
-	ui_callback_fn_t* this_fn = signTx_handleScriptDataHash_ui_runStep;
-
-	UI_STEP_BEGIN(ctx->ui_step, this_fn);
-
-	UI_STEP(HANDLE_SCRIPT_DATA_HASH_STEP_DISPLAY) {
-		ui_displayBech32Screen(
-		        "Script data hash",
-		        "script_data",
-		        BODY_CTX->stageData.scriptDataHash, SCRIPT_DATA_HASH_LENGTH,
-		        this_fn
-		);
-	}
-	UI_STEP(HANDLE_SCRIPT_DATA_HASH_STEP_RESPOND) {
-		respondSuccessEmptyMsg();
-		advanceStage();
-	}
-	UI_STEP_END(HANDLE_FEE_STEP_INVALID);
-}
-
 static void signTx_handleScriptDataHashAPDU(uint8_t p2, const uint8_t* wireDataBuffer, size_t wireDataSize)
 {
 	{
@@ -1938,7 +1392,7 @@ static void ui_advanceState_collateralInput()
 	BODY_CTX->currentCollateral++;
 
 	if (BODY_CTX->currentCollateral == ctx->numCollateralInputs) {
-		advanceStage();
+		tx_advanceStage();
 	}
 }
 
@@ -1983,54 +1437,6 @@ static void signTx_handleCollateralInputAPDU(uint8_t p2, const uint8_t* wireData
 }
 
 // ========================= REQUIRED SIGNERS ===========================
-
-enum {
-	HANDLE_REQUIRED_SIGNERS_STEP_DISPLAY = 1400,
-	HANDLE_REQUIRED_SIGNERS_STEP_RESPOND,
-	HANDLE_REQUIRED_SIGNERS_STEP_INVALID,
-};
-
-static void signTx_handleRequiredSigner_ui_runStep()
-{
-	TRACE("UI step %d", ctx->ui_step);
-	ui_callback_fn_t* this_fn = signTx_handleRequiredSigner_ui_runStep;
-
-	UI_STEP_BEGIN(ctx->ui_step, this_fn);
-
-	UI_STEP(HANDLE_REQUIRED_SIGNERS_STEP_DISPLAY) {
-		switch (BODY_CTX->stageData.requiredSigner.type) {
-		case REQUIRED_SIGNER_WITH_PATH:
-			ui_displayPathScreen("Required signer", &BODY_CTX->stageData.requiredSigner.keyPath, this_fn);
-			break;
-		case REQUIRED_SIGNER_WITH_HASH:
-			ui_displayBech32Screen(
-			        "Required signer",
-			        "req_signer_vkh",
-			        BODY_CTX->stageData.requiredSigner.keyHash,
-			        SIZEOF(BODY_CTX->stageData.requiredSigner.keyHash),
-			        this_fn
-			);
-			break;
-
-		default:
-			ASSERT(false);
-			break;
-		}
-	}
-
-	UI_STEP(HANDLE_REQUIRED_SIGNERS_STEP_RESPOND) {
-		respondSuccessEmptyMsg();
-
-		// Advance stage to the next input
-		ASSERT(BODY_CTX->currentRequiredSigner < ctx->numRequiredSigners);
-		BODY_CTX->currentRequiredSigner++;
-
-		if (BODY_CTX->currentRequiredSigner == ctx->numRequiredSigners) {
-			advanceStage();
-		}
-	}
-	UI_STEP_END(HANDLE_REQUIRED_SIGNERS_STEP_INVALID);
-}
 
 __noinline_due_to_stack__
 static void signTx_handleRequiredSignerAPDU(uint8_t p2, const uint8_t* wireDataBuffer, size_t wireDataSize)
@@ -2125,31 +1531,6 @@ static void signTx_handleCollateralOutputAPDU(uint8_t p2, const uint8_t* wireDat
 
 // ========================= TOTAL COLLATERAL ===========================
 
-enum {
-	HANDLE_TOTAL_COLLATERAL_STEP_DISPLAY = 400,
-	HANDLE_TOTAL_COLLATERAL_STEP_RESPOND,
-	HANDLE_TOTAL_COLLATERAL_STEP_INVALID,
-};
-
-static void signTx_handleTotalCollateral_ui_runStep()
-{
-	TRACE("UI step %d", ctx->ui_step);
-	ui_callback_fn_t* this_fn = signTx_handleTotalCollateral_ui_runStep;
-
-	TRACE_ADA_AMOUNT("total collateral ", BODY_CTX->stageData.totalCollateral);
-
-	UI_STEP_BEGIN(ctx->ui_step, this_fn);
-
-	UI_STEP(HANDLE_TOTAL_COLLATERAL_STEP_DISPLAY) {
-		ui_displayAdaAmountScreen("Total collateral", BODY_CTX->stageData.totalCollateral, this_fn);
-	}
-	UI_STEP(HANDLE_TOTAL_COLLATERAL_STEP_RESPOND) {
-		respondSuccessEmptyMsg();
-		advanceStage();
-	}
-	UI_STEP_END(HANDLE_TOTAL_COLLATERAL_STEP_INVALID);
-}
-
 __noinline_due_to_stack__
 static void signTx_handleTotalCollateralAPDU(uint8_t p2, const uint8_t* wireDataBuffer, size_t wireDataSize)
 {
@@ -2206,7 +1587,7 @@ static void ui_advanceState_ReferenceInput()
 	BODY_CTX->currentReferenceInput++;
 
 	if (BODY_CTX->currentReferenceInput == ctx->numReferenceInputs) {
-		advanceStage();
+		tx_advanceStage();
 	}
 }
 
@@ -2248,45 +1629,6 @@ static void signTx_handleReferenceInputsAPDU(uint8_t p2, const uint8_t* wireData
 }
 
 // ============================== CONFIRM ==============================
-
-enum {
-	HANDLE_CONFIRM_STEP_TXID = 1000,
-	HANDLE_CONFIRM_STEP_FINAL_CONFIRM,
-	HANDLE_CONFIRM_STEP_RESPOND,
-	HANDLE_CONFIRM_STEP_INVALID,
-};
-
-static void signTx_handleConfirm_ui_runStep()
-{
-	TRACE("UI step %d", ctx->ui_step);
-	TRACE_STACK_USAGE();
-	ui_callback_fn_t* this_fn = signTx_handleConfirm_ui_runStep;
-
-	UI_STEP_BEGIN(ctx->ui_step, this_fn);
-
-	UI_STEP(HANDLE_CONFIRM_STEP_TXID) {
-		ui_displayHexBufferScreen(
-		        "Transaction id",
-		        ctx->txHash, SIZEOF(ctx->txHash),
-		        this_fn
-		);
-	}
-	UI_STEP(HANDLE_CONFIRM_STEP_FINAL_CONFIRM) {
-		ui_displayPrompt(
-		        "Confirm",
-		        "transaction?",
-		        this_fn,
-		        respond_with_user_reject
-		);
-	}
-	UI_STEP(HANDLE_CONFIRM_STEP_RESPOND) {
-		io_send_buf(SUCCESS, ctx->txHash, SIZEOF(ctx->txHash));
-		ui_displayBusy(); // displays dots, called only after I/O to avoid freezing
-
-		advanceStage();
-	}
-	UI_STEP_END(HANDLE_CONFIRM_STEP_INVALID);
-}
 
 static bool _shouldDisplayTxId(sign_tx_signingmode_t signingMode)
 {
@@ -2355,61 +1697,6 @@ static void signTx_handleConfirmAPDU(uint8_t p2, const uint8_t* wireDataBuffer M
 
 
 // ============================== WITNESS ==============================
-
-static void _wipeWitnessSignature()
-{
-	// safer not to keep the signature in memory
-	explicit_bzero(WITNESS_CTX->stageData.witness.signature, SIZEOF(WITNESS_CTX->stageData.witness.signature));
-	respond_with_user_reject();
-}
-
-enum {
-	HANDLE_WITNESS_STEP_WARNING = 1100,
-	HANDLE_WITNESS_STEP_DISPLAY,
-	HANDLE_WITNESS_STEP_CONFIRM,
-	HANDLE_WITNESS_STEP_RESPOND,
-	HANDLE_WITNESS_STEP_INVALID,
-};
-
-static void signTx_handleWitness_ui_runStep()
-{
-	TRACE("UI step %d", ctx->ui_step);
-	TRACE_STACK_USAGE();
-	ui_callback_fn_t* this_fn = signTx_handleWitness_ui_runStep;
-
-	UI_STEP_BEGIN(ctx->ui_step, this_fn);
-
-	UI_STEP(HANDLE_WITNESS_STEP_WARNING) {
-		ui_displayPaginatedText(
-		        "WARNING:",
-		        "unusual witness requested",
-		        this_fn
-		);
-	}
-	UI_STEP(HANDLE_WITNESS_STEP_DISPLAY) {
-		ui_displayPathScreen("Witness path", &WITNESS_CTX->stageData.witness.path, this_fn);
-	}
-	UI_STEP(HANDLE_WITNESS_STEP_CONFIRM) {
-		ui_displayPrompt(
-		        "Sign using",
-		        "this witness?",
-		        this_fn,
-		        _wipeWitnessSignature
-		);
-	}
-	UI_STEP(HANDLE_WITNESS_STEP_RESPOND) {
-		TRACE("Sending witness data");
-		TRACE_BUFFER(WITNESS_CTX->stageData.witness.signature, SIZEOF(WITNESS_CTX->stageData.witness.signature));
-		io_send_buf(SUCCESS, WITNESS_CTX->stageData.witness.signature, SIZEOF(WITNESS_CTX->stageData.witness.signature));
-		ui_displayBusy(); // displays dots, called only after I/O to avoid freezing
-
-		WITNESS_CTX->currentWitness++;
-		if (WITNESS_CTX->currentWitness == ctx->numWitnesses) {
-			advanceStage();
-		}
-	}
-	UI_STEP_END(HANDLE_WITNESS_STEP_INVALID);
-}
 
 __noinline_due_to_stack__
 static void signTx_handleWitnessAPDU(uint8_t p2, const uint8_t* wireDataBuffer, size_t wireDataSize)
